worst code, but working

+++ db/migrate/20150223112217_create_positions.rb
@@ -2,9 +2,9 @@ class CreatePositions < ActiveRecord::Migration
   def change
     create_table :positions do |t|
     	t.belongs_to :team
-    	t.integer :latitude
-    	t.integer :longitude
-    	t.integer :distance 
+    	t.float :latitude
+    	t.float :longitude
+    	t.float :distance 
     	t.string :text
       t.timestamps null: false
     end
